requirements.txt
Flask==3.0.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
pymysql==1.1.0
requests==2.32.3
gunicorn==23.0.0
